extract abstract restart test
